requirements.txt
doit
requests
pycdlib
python-dotenv
